serve: use epoll
The `serve` program now uses epoll so that it does not get blocked
on connections that are kept open, but do not necessarily send their
requests for some time.  Such connections are actually made by real web
browsers when running multiple `fetch()` functions (tested with Firefox
and Chromium on Arch Linux), so that behavior made the server unusable
for such scenarios.

However, this commit makes the server exclusive to Linux.  It might
actually be a better idea to use `poll()` for this?